feat/find 서버 클라이언트 분리,라벨,메타데이터추가
